Remove options and background, remove mantine
