--- .config/options.config.ts
@@ -1,17 +1,0 @@
-import { defineConfig, mergeConfig } from 'vite'
-import baseConfig from '../vite.config.base'
-
-// https://vitejs.dev/config/
-export default mergeConfig(
-  baseConfig,
-  defineConfig({
-  build: {
-    outDir: "build/options/",
-    rollupOptions: {
-        input: {
-            options: "options.html"
-        }
-    }
-  }
-})
-)
