--- src/features/options/OptionsPage/OptionsPage.module.css
@@ -1,42 +1,0 @@
-/* Please tell me if this is a backwards way of doing things. */
-:root {
-    --background-1: #1F1F1F;
-    --text-1: #46AC99;
-    --text-2: #CCCCCC;
-
-}
-
-.header {
-    & > div {
-        height: 100%;
-        justify-content: center;
-        padding: 1rem;
-        & > img {
-            object-fit: contain;
-            height: 100%;
-        }
-    }
-}
-
-.main {
-    display: flex;
-    justify-content: center;
-    & > div {
-        width: 100%;
-        display: flex;
-        justify-content: center;
-        /* Stack */
-        & > div { 
-            max-width: 50rem;
-            width: 100%;
-            padding: 2rem 5rem;
-            gap: 2rem;
-            border-left: solid 1px var(--mantine-color-dark-outline);
-            border-right: solid 1px var(--mantine-color-dark-outline);
-            
-            & > button {
-                width: fit-content;
-            }
-        }
-    }
-}
