--- src/features/options/index.css
@@ -1,8 +1,0 @@
-* {
-    box-sizing: border-box;
-}
-
-#root {
-    width: 100%;
-    height: 100%;
-}
