+++ src/features/popup/index.tsx
@@ -2,14 +2,10 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
-import { MantineProvider } from '@mantine/core'
-import '@mantine/core/styles.css';
 
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <MantineProvider defaultColorScheme='dark'>
       <App />
-    </MantineProvider>
   </StrictMode>,
 )
